build: divide comments adn reports into 2 components
