+++ admin/src/pages/DashboardPage/styles.ts
@@ -1,0 +1,12 @@
+//@ts-nocheck
+import styled from "styled-components";
+import { Box } from "@strapi/design-system/Box";
+
+export const PanelLayout = styled(Box)`
+    margin-top: 10px;
+    padding-top: 10px;
+`;
+
+export const StyledHeader = styled(Box)`
+    margin-bottom: 10px;
+`;
